jnvm-18: users cant use versions they don't have installed

+++ src/commands/use.ts
@@ -2,10 +2,14 @@ import fs from "fs";
 import path from "node:path";
 
 export async function useVersion(nodeVersion: string, symLinkPath: string, jnvmPath: string) {
+    if (!fs.existsSync(path.join(jnvmPath, nodeVersion))){
+        return false;
+    }
     try {
         await fs.promises.rmdir(symLinkPath);
     }catch (e) {
         console.log('No symlink directory to remove');
     }
-    return fs.promises.symlink(path.join(jnvmPath, nodeVersion), symLinkPath, 'dir');
+    await fs.promises.symlink(path.join(jnvmPath, nodeVersion), symLinkPath, 'dir');
+    return true;
 }
